chore: apply setSelectionByUI util
